refactor: move layout config from app to layout classes
