+++ static/js/layout/layout.embed.js
@@ -6,6 +6,38 @@ export default class EmbedLayout extends Layout {
     'Click the "Run" button to execute code.',
   ];
 
+  /**
+   * Create the layout.
+   *
+   * @param {array} content - List of content objects.
+   * @param {number} fontSize - The default font-size to be used.
+   * @param {object} options - Additional options object.
+   * @param {boolean} options.vertical - Whether the layout should be vertical.
+   * @param {string} options.proglang - The programming language to be used
+   */
+  constructor(content, fontSize, options = {}) {
+    const defaultLayoutConfig = {
+      dimensions: {
+        borderWidth: 0,
+      },
+      content: [
+        {
+          type: options.vertical ? 'column' : 'row',
+          content: [
+            {
+              content,
+            },
+            {
+              componentState: { fontSize },
+            }
+          ]
+        }
+      ]
+    };
+
+    super(defaultLayoutConfig, options);
+  }
+
   renderButtons() {
     const runCodeButtonHtml = this.getRunCodeButtonHtml();
     const settingsMenuHtml = this.getSettingsMenuHtml();
